+++ envdata/urls.py
@@ -82,6 +82,7 @@ urlpatterns = [
     path('travel-pdf-report', views.TravelPdfView.as_view(), name='travel_report'),
 
     path('upload-fuel-data', views.ExcelUploadView.as_view(), name='upload_fuel_data'),
-    path('upload-ng-data', views.ExcelUploadViewForNaturalGas.as_view(), name='upload_ng_data')
+    path('upload-ng-data', views.ExcelUploadViewForNaturalGas.as_view(), name='upload_ng_data'),
+    path('upload-energy-data',views.ExcelUploadViewForEnergy.as_view(),name='upload_energy_data')
 
 ]